+++ scripts/lib/runtests.js
@@ -7,20 +7,33 @@ var utils = require('./utils');
 var lookup = utils.lookup;
 var promiseSequence = utils.promiseSequence;
 
-function mochaRun() {
+function mochaRun({cliTest = false} = {}) {
+  // We need to run the cli test without nyc because of weird behavior
+  // with spawn-wrap
+  const bin = lookup((cliTest) ? '.bin/mocha' : '.bin/nyc', true);
+  const runArgs = (cliTest)
+    ? []
+    : [
+      '--require', '@babel/register',
+      '--exclude',
+      'tests/**',
+      '--silent',
+      '--no-clean',
+      require.resolve('mocha/bin/mocha'),
+    ];
+
+  const mochaArgs = (cliTest)
+    ? ['tests/cli.js']
+    : ['--grep', 'precompile cli', '--invert', 'tests'];
+
   return new Promise((resolve, reject) => {
     try {
-      const proc = spawn(lookup('.bin/nyc', true), [
-        '--require', '@babel/register',
-        '--exclude',
-        'tests/**',
-        '--silent',
-        '--no-clean',
-        require.resolve('mocha/bin/mocha'),
+      const proc = spawn(bin, [
+        ...runArgs,
         '-R', 'spec',
         '-r', 'tests/setup',
         '-r', '@babel/register',
-        'tests'
+        ...mochaArgs,
       ], {
         cwd: path.join(__dirname, '../..'),
         env: process.env
@@ -47,7 +60,13 @@ function mochaRun() {
 function runtests() {
   return new Promise((resolve, reject) => {
     var server;
-    return mochaRun().then(() => {
+
+    const mochaPromise = promiseSequence([
+      () => mochaRun({cliTest: false}),
+      () => mochaRun({cliTest: true}),
+    ]);
+
+    return mochaPromise.then(() => {
       return getStaticServer().then((args) => {
         server = args[0];
         const port = args[1];
@@ -68,4 +87,3 @@ function runtests() {
 }
 
 module.exports = runtests;
-